Fix trusted certs valid? method so it returns false for non-existant files

+++ files/gitlab-cookbooks/gitlab/libraries/helper.rb
@@ -469,12 +469,10 @@ class CertificateHelper
   end
 
   def valid?(file)
-    if File.symlink?(file) && !File.exists?(file)
-      FileUtils.rm_f(file)
-      return false
-    end
+    exists = File.exists?(file)
+    FileUtils.rm_f(file) if File.symlink?(file) && !exists
 
-    true
+    exists
   end
 
   def move_certificate(file)
@@ -482,14 +480,15 @@ class CertificateHelper
 
     # Move the certs to the trusted certs directory if it is located within our managed certs directory
     # Otherwise copy the cert to the trusted certs directory
-    if File.realpath(file).start_with?(@omnibus_certs_dir)
-      FileUtils.mv(File.realpath(file), @trusted_certs_dir, force: true)
+    realpath = File.realpath(file)
+    if realpath.start_with?(@omnibus_certs_dir)
+      FileUtils.mv(realpath, @trusted_certs_dir, force: true)
     else
-      FileUtils.cp(File.realpath(file), @trusted_certs_dir)
+      FileUtils.cp(realpath, @trusted_certs_dir)
     end
 
     FileUtils.rm_f(file) if File.symlink?(file)
-    puts "\n Moving #{File.realpath(file)}"
+    puts "\n Moving #{realpath}"
   end
 
   def link_certificates
